feat: Add film-list controller
